+++ source/chip/STM32/STM32F1/Tupfile.lua
@@ -9,10 +9,13 @@
 
 if CONFIG_CHIP_STM32F1 == "y" then
 
-	local stm32f1LdSh = "STM32F1.ld.sh"
+	local ldScriptGenerator = TOP .. "source/architecture/ARM/ARMv7-M/ARMv7-M.ld.sh"
 	local ldscriptOutputs = {LDSCRIPT, filenameToGroup(LDSCRIPT)}
 
-	tup.rule(DISTORTOS_CONFIGURATION_MK, "^ SH " .. stm32f1LdSh .. "^ ./" .. stm32f1LdSh .. " \"%f\" > \"%o\"",
+	tup.rule("^ SH " .. ldScriptGenerator .. "^ ./" .. ldScriptGenerator .. " \"" .. CONFIG_CHIP .. "\" \"" ..
+			CONFIG_CHIP_STM32F1_FLASH_ADDRESS .. "," .. CONFIG_CHIP_STM32F1_FLASH_SIZE .. "\" \"" ..
+			CONFIG_CHIP_STM32F1_SRAM_ADDRESS .. "," .. CONFIG_CHIP_STM32F1_SRAM_SIZE .. "\" \"" ..
+			CONFIG_ARCHITECTURE_ARMV7_M_MAIN_STACK_SIZE .. "\" \"" .. CONFIG_MAIN_THREAD_STACK_SIZE .. "\" > \"%o\"",
 			ldscriptOutputs)
 
 	CFLAGS += STANDARD_INCLUDES